+++ src/main/java/cburgosdev/java/Repositories/IProductRecordRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface IProductRecordRepository extends JpaRepository<ProductRecord, Long> {
-    List<ProductRecord> getProductRecordListByProductId(Sort sort, int productId);
+    List<ProductRecord> getProductRecordListByProductId(Sort sort, Long productId);
 }
